3d geometry test

+++ src/com/zn0w/zfsr3d/math/Matrix.java
@@ -2,14 +2,14 @@ package com.zn0w.zfsr3d.math;
 
 public class Matrix {
 	
-	public int[][] values;
+	public double[][] values;
 	
 	
 	public Matrix(int size) {
-		values = new int[size][size];
+		values = new double[size][size];
 	}
 	
-	public Matrix(int[][] values) {
+	public Matrix(double[][] values) {
 		this.values = values;
 	}
